authentication & multer

+++ controllers/stuff.js
@@ -3,12 +3,14 @@
 const Thing = require('../models/thing');
 
 exports.createThing = (req, res, next) => {
+  req.body.thing = JSON.parse(req.body.thing);
+  const url = req.protocol + '://' + req.get('host');
   const thing = new Thing({
-    title: req.body.title,
-    description: req.body.description,
-    imageUrl: req.body.imageUrl,
-    price: req.body.price,
-    userId: req.body.userId
+    title: req.body.thing.title,
+    description: req.body.thing.description,
+    imageUrl: url + '/images/' + req.file.filename,
+    price: req.body.thing.price,
+    userId: req.body.thing.userId
   });
   thing.save().then(
     () => {
@@ -42,14 +44,28 @@ exports.getOneThing = (req, res, next) => {
 };
 
 exports.modifyThing = (req, res, next) => {
-  const thing = new Thing({
-    _id: req.params.id,
-    title: req.body.title,
-    description: req.body.description,
-    imageUrl: req.body.imageUrl,
-    price: req.body.price,
-    userId: req.body.userId
-  });
+  let thing = new Thing({ _id: req.params._id });
+  if (req.file) {
+    const url = req.protocol + '://' + req.get('host');
+    req.body.thing = JSON.parse(req.body.thing);
+    thing = {
+      _id: req.params.id,
+      title: req.body.thing.title,
+      description: req.body.thing.description,
+      imageUrl: url + '/images/' + req.file.filename,
+      price: req.body.thing.price,
+      userId: req.body.thing.userId
+    };
+  } else {
+    thing = {
+      _id: req.params.id,
+      title: req.body.title,
+      description: req.body.description,
+      imageUrl: req.body.imageUrl,
+      price: req.body.price,
+      userId: req.body.userId
+    };
+  }
   Thing.updateOne({_id: req.params.id}, thing).then(
     () => {
       res.status(201).json({
@@ -66,22 +82,36 @@ exports.modifyThing = (req, res, next) => {
 };
 
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({_id: req.params.id}).then(
-    () => {
-      res.status(200).json({
-        message: 'Deleted!'
-      });
-    }
-  ).catch(
-    (error) => {
-      res.status(400).json({
-        error: error
-      });
+  Thing.findOne({ _id: req.params.id }).then(
+    (thing) => {
+      if (!thing) {
+        return res.status(404).json({
+          error: new Error('Objet non trouvé !')
+        });
+      }
+      if (thing.userId !== req.auth.userId) {
+        return res.status(401).json({
+          error: new Error('Request Not Authorized !')
+        });
+      }
+      Thing.deleteOne({_id: req.params.id}).then(
+        () => {
+          res.status(200).json({
+            message: 'Deleted!'
+          });
+        }
+      ).catch(
+        (error) => {
+          res.status(400).json({
+            error: error
+          });
+        }
+      );
     }
   );
 };
 
-exports.getAllStuff = (req, res, next) => {
+exports.getAllStuff =  (req, res, next) => {
   Thing.find().then(
     (things) => {
       res.status(200).json(things);
